Create and populate `plasmapy.formulary.speeds` (#1448)
