+++ changelog/1448.feature.rst
@@ -1,0 +1,10 @@
+Created `plasmapy.formulary.speeds` to contain frequency related
+plasma parameters, and migrated
+`~plasmapy.formulary.speeds.Alfven_speed`,
+`~plasmapy.formulary.speeds.ion_sound_speed`,
+`~plasmapy.formulary.speeds.kappa_thermal_speed`,
+`~plasmapy.formulary.speeds.thermal_speed`,
+`~plasmapy.formulary.speeds.thermal_speed_coefficients`, and
+`~plasmapy.formulary.speeds.thermal_speed_lite` from
+`plasmapy.formulary.parameters` to the new module.  Related aliases were
+also migrated.
