+++ docs/glossary.rst
@@ -57,12 +57,12 @@ Glossary
       - Lite-functions are a flavor of existing `plasmapy`
         functionality and are denoted with a trailing ``_lite`` to the
         name. For example,
-        `~plasmapy.formulary.parameters.thermal_speed_lite` is the
+        `~plasmapy.formulary.speeds.thermal_speed_lite` is the
         lite-function associated with
-        `~plasmapy.formulary.parameters.thermal_speed`.
+        `~plasmapy.formulary.speeds.thermal_speed`.
       - Lite-functions are typically bound to their normal version as
         the ``lite`` attribute.  For example,
-        `~plasmapy.formulary.parameters.thermal_speed_lite` can be
+        `~plasmapy.formulary.speeds.thermal_speed_lite` can be
         accessed as ``thermal_speed.lite``.
       - Lite-functions view all input and output arguments as being in
         SI units.
