+++ plasmapy/dispersion/analytical/tests/test_two_fluid_.py
@@ -6,8 +6,8 @@ import pytest
 from astropy import units as u
 
 from plasmapy.dispersion.analytical.two_fluid_ import two_fluid
-from plasmapy.formulary import parameters as pfp
 from plasmapy.formulary.frequencies import wc_
+from plasmapy.formulary.speeds import cs_, va_
 from plasmapy.particles import Particle
 from plasmapy.utils.exceptions import PhysicsWarning
 
@@ -116,8 +116,8 @@ class TestTwoFluid:
         # theta and k values need to be single valued for this test to function
         # correctly
 
-        cs = pfp.cs_(kwargs["T_e"], kwargs["T_i"], kwargs["ion"])
-        va = pfp.va_(kwargs["B"], kwargs["n_i"], ion=kwargs["ion"])
+        cs = cs_(kwargs["T_e"], kwargs["T_i"], kwargs["ion"])
+        va = va_(kwargs["B"], kwargs["n_i"], ion=kwargs["ion"])
         wci = wc_(kwargs["B"], kwargs["ion"])
 
         beta = (cs / va).value ** 2
