+++ plasmapy/dispersion/numerical/hollweg_.py
@@ -10,8 +10,8 @@ import warnings
 from astropy.constants.si import c
 from typing import Union
 
-from plasmapy.formulary import parameters as pfp
 from plasmapy.formulary.frequencies import gyrofrequency, plasma_frequency
+from plasmapy.formulary.speeds import Alfven_speed, ion_sound_speed
 from plasmapy.particles import Particle
 from plasmapy.particles.exceptions import ChargeError
 from plasmapy.utils.decorators import validate_quantities
@@ -273,7 +273,7 @@ def hollweg(
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", category=PhysicsWarning)
         n_e = z_mean * n_i
-        c_s = pfp.ion_sound_speed(
+        c_s = ion_sound_speed(
             T_e=T_e,
             T_i=T_i,
             ion=ion,
@@ -282,7 +282,7 @@ def hollweg(
             gamma_i=gamma_i,
             z_mean=z_mean,
         ).value
-        v_A = pfp.Alfven_speed(B, n_i, ion=ion, z_mean=z_mean).value
+        v_A = Alfven_speed(B, n_i, ion=ion, z_mean=z_mean).value
         omega_ci = gyrofrequency(B=B, particle=ion, signed=False, Z=z_mean).value
         omega_pe = plasma_frequency(n=n_e, particle="e-").value
